Added numreffiles and numqueryfiles parameters
Memory usage control added:
Added the ability to specify how many parts the reference and query
files for a MUMmer run should be split into (defaults: 10 for ref and
10 for query, as is in the original code).  Since, according to Adam
Phillippy, the suffix tree in MUMmer uses about 17 bytes per base of
the reference, having nProcs*17*refpartlength bytes being used may
exceed the available memory.

I have experienced this with two datasets (Drosophila yakuba where the
reference was ~170 Mb, and Papilio glaucus where the reference was
350-500 Mb), which leads to failure at or before generation of
improved.fasta on a server with 380 GB RAM.  The combined MUMmer jobs
request between 500 GB and 1 TB of RAM for D. yakuba (for -par 20 and
-par 10, respectively), so this allows for extra splitting of the query
without allocating extra processes from the Pool, and extra splitting
of the reference as a parallel route for reducing memory consumption at
the expense of runtime.

Also, is the "combineMultipleCoorMum" method defunct?  I couldn't find
it called anywhere, and the use of head -n5 and tail -n+6 appear
elsewhere.