Fix for MUMming a file against itself
With the addition of independently varying the number of parts the
reference and query are split into, MUMming a file against itself is
now broken, since the file splitting overwrites its own files.

My fix is to create a symlink to the input file if the reference and
query files are the same, and allow the splitter script to split using
the filename of the symlink, thereby preventing overwriting the
reference split files.

Also added some notes about implementing working directory cleanup in
the future.