Added many print statements to keep track of position in the workflow
Added print statements to describe in the logs where we are in the
workflow.  I didn't quite grasp the purpose of some sections in
gapFiller.py and twoRepeatOneBridgeSolver.py, so some sections are not
annotated.  However, this should help with diagnostics of which part
failed, and what major output files should be seen or examined.