Attempt #1 to make FinisherSC use SLURM arrays
Edits to attempt to make FinisherSC use SLURM arrays for its MUMmer
jobs, thereby further parallelizing it.  Currently, it is only
sometimes possible to get FinisherSC to complete for a 170 Mb genome
within 1 day given 20 cores and 380 GB RAM, but this might allow for
using more RAM in total, but split across cores.

Code changes include the use of file locks to perform blocking and
synchronization of the main tasks, and the creation of a template
script for submission as a SLURM task array script.