More lock directory fixes...
For some reason, os.mkdir doesn't seem to sync before the SLURM jobs
run, so touch fails because the locks directory doesn't exist when it
tries to create a lock file.  In it's stead, now the SLURM jobs will
create the directory (and not clobber, due to mkdir -p silently failing
if directory exists), thus guaranteeing existence prior to touching.