Added polling to cope with filesystem latency
In order to cope with slow updates across nodes for NFS, we have to
poll for the number of files in the locks directory until all locks are
in place.  Then, and only then, can we iterate through the list of
locks reliably, and wait for the MUMmer jobs to finish.